refactor(configfile): rename configparse
