+++ internal/configparse/error.go
@@ -1,4 +1,4 @@
-package configfile
+package configparse
 
 import (
 	"errors"
