+++ internal/configparse/testdata/extends/extends-circular-0.yml
