+++ internal/configparse/testdata/extends/extends-circular-1.yml
