+++ internal/configparse/testdata/extends/extends-circular-2.yml
