+++ internal/configparse/testdata/extends/extends-circular-self.yml
