+++ internal/configparse/testdata/extends/extends-valid-child.yml
