+++ internal/configparse/testdata/extends/extends-valid-parent.yml
