+++ internal/configparse/testdata/extends/nest-0/nest-1/extends-valid-nested.yml
