+++ internal/configparse/testdata/invalid/badfields.json
