+++ internal/configparse/testdata/invalid/badfields.yml
