+++ internal/configparse/testdata/valid/benchttp.json
