+++ internal/configparse/testdata/valid/benchttp.yaml
