+++ internal/configparse/testdata/valid/benchttp.yml
